Update table filter config
