+++ data-prepper-plugins/rds-source/src/main/java/org/opensearch/dataprepper/plugins/source/rds/RdsService.java
@@ -14,7 +14,6 @@ import org.opensearch.dataprepper.model.event.EventFactory;
 import org.opensearch.dataprepper.model.plugin.PluginConfigObservable;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.model.source.coordinator.enhanced.EnhancedSourceCoordinator;
-import org.opensearch.dataprepper.plugins.source.rds.configuration.TableFilterConfig;
 import org.opensearch.dataprepper.plugins.source.rds.export.DataFileScheduler;
 import org.opensearch.dataprepper.plugins.source.rds.export.ExportScheduler;
 import org.opensearch.dataprepper.plugins.source.rds.export.ExportTaskManager;
@@ -213,9 +212,8 @@ public class RdsService {
     }
 
     private Map<String, Map<String, String>> getColumnDataTypeMap(final SchemaManager schemaManager) {
-        TableFilterConfig tableFilterConfig = sourceConfig.getTables();
-        Set<String> tableNames = schemaManager.getTableNames(tableFilterConfig.getDatabase());
-        tableFilterConfig.applyTableFilter(tableNames);
+        Set<String> tableNames = schemaManager.getTableNames(sourceConfig.getDatabase());
+        sourceConfig.applyTableFilter(tableNames);
         LOG.info("These tables will be include in processing: {}", tableNames);
         return schemaManager.getColumnDataTypes(new ArrayList<>(tableNames));
     }
