+++ data-prepper-plugins/rds-source/src/main/java/org/opensearch/dataprepper/plugins/source/rds/RdsSourceConfig.java
@@ -10,6 +10,7 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import org.opensearch.dataprepper.plugins.source.rds.configuration.AwsAuthenticationConfig;
 import org.opensearch.dataprepper.plugins.source.rds.configuration.EngineType;
@@ -19,6 +20,9 @@ import org.opensearch.dataprepper.plugins.source.rds.configuration.TlsConfig;
 import software.amazon.awssdk.regions.Region;
 
 import java.time.Duration;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * Configuration for RDS Source
@@ -42,6 +46,10 @@ public class RdsSourceConfig {
     @JsonProperty("engine")
     @NotNull
     private EngineType engine;
+
+    @JsonProperty("database")
+    @NotEmpty
+    private String database;
 
     @JsonProperty("tables")
     private TableFilterConfig tableFilterConfig;
@@ -109,6 +117,10 @@ public class RdsSourceConfig {
 
     public boolean isAurora() {
         return engine.isAurora();
+    }
+
+    public String getDatabase() {
+        return database;
     }
 
     public TableFilterConfig getTables() {
@@ -188,6 +200,27 @@ public class RdsSourceConfig {
 
         public String getPassword() {
             return password;
+        }
+    }
+
+    /**
+     * This method applies the table filter configuration to the given set of table names.
+     *
+     * @param tableNames        The set of table names to be filtered
+     */
+    public void applyTableFilter(Set<String> tableNames) {
+        if (!tableFilterConfig.getInclude().isEmpty()) {
+            List<String> includeTableList = tableFilterConfig.getInclude().stream()
+                    .map(item -> getDatabase() + "." + item)
+                    .collect(Collectors.toList());
+            tableNames.retainAll(includeTableList);
+        }
+
+        if (!tableFilterConfig.getExclude().isEmpty()) {
+            List<String> excludeTableList = tableFilterConfig.getExclude().stream()
+                    .map(item -> getDatabase() + "." + item)
+                    .collect(Collectors.toList());
+            excludeTableList.forEach(tableNames::remove);
         }
     }
 }
