+++ data-prepper-plugins/rds-source/src/test/java/org/opensearch/dataprepper/plugins/source/rds/RdsServiceTest.java
@@ -23,7 +23,6 @@ import org.opensearch.dataprepper.model.plugin.PluginConfigObservable;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.model.source.coordinator.enhanced.EnhancedSourceCoordinator;
 import org.opensearch.dataprepper.plugins.source.rds.configuration.EngineType;
-import org.opensearch.dataprepper.plugins.source.rds.configuration.TableFilterConfig;
 import org.opensearch.dataprepper.plugins.source.rds.configuration.TlsConfig;
 import org.opensearch.dataprepper.plugins.source.rds.export.DataFileScheduler;
 import org.opensearch.dataprepper.plugins.source.rds.export.ExportScheduler;
@@ -196,13 +195,11 @@ class RdsServiceTest {
                                 .build())
                         .build())
                 .build();
-        final TableFilterConfig tableFilterConfig = mock(TableFilterConfig.class);
         final String databaseName = UUID.randomUUID().toString();
         final Set<String> tableNames = Set.of("database1.table1", "database2.table2");
 
         when(sourceConfig.getDbIdentifier()).thenReturn(dbIdentifier);
-        when(sourceConfig.getTables()).thenReturn(tableFilterConfig);
-        when(tableFilterConfig.getDatabase()).thenReturn(databaseName);
+        when(sourceConfig.getDatabase()).thenReturn(databaseName);
         when(schemaManager.getTableNames(databaseName)).thenReturn(tableNames);
         when(schemaManager.getColumnDataTypes(new ArrayList<>(tableNames))).thenReturn(mock(Map.class));
         when(rdsClient.describeDBInstances(any(DescribeDbInstancesRequest.class))).thenReturn(describeDbInstancesResponse);
